requirements.txt
biopython==1.85
pandas==2.3.3
matplotlib==3.10.7
numpy==2.3.4

protein-mutation-analyzer/
├── README.md
├── .gitignore
├── requirements.txt
├── data/
├── scripts/
└── output/